% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_calculate-returns.R
\name{calc_ret_track}
\alias{calc_ret_track}
\title{Calculate Returns (Single)}
\usage{
calc_ret_track(t, dist, lag, window = Inf)
}
\arguments{
\item{dist}{Numeric. Maximum distance (UNIT...m?) to be considered a return}

\item{lag}{Vector of fidelity lags (one or more)}

\item{window}{# Duration of the fidelity window (Inf if from the start of the track, an integer if shorter (e.g., 42 locations, which is a week at 6h resolution). Default is Inf.}

\item{tracks}{Data frame of a movement track (e.g., output from simulate_tracks(), or a set of real tracks) with at minimum columns for id, x, y, and step.}
}
\value{
Data frame of a movement track containing an additional column indicating a return (1 = return, 0 = no return) for each lag.
}
\description{
\code{calc_ret_track} Function to calculate returns for a single track based on temporal and spatial restrictions.
}
