% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_calculate-returns.R
\name{calc_returns}
\alias{calc_returns}
\title{Calculate Returns (Multiple)}
\usage{
calc_returns(tracks, dist, lag, window = Inf)
}
\arguments{
\item{tracks}{Data frame of movement tracks (e.g., output from simulate_tracks(), or a set of real tracks) with at minimum columns for id, x, y, and step.}

\item{dist}{Numeric. Maximum distance (UNIT...m?) to be considered a return}

\item{lag}{Vector of fidelity lags (one or more)}

\item{window}{# Duration of the fidelity window (Inf if from the start of the track, an integer if shorter (e.g., 42 locations, which is a week at 6h resolution). Default is Inf.}
}
\value{
Data frame of movement tracks containing an additional column indicating a return (1 = return, 0 = no return) for each lag.
}
\description{
\code{calc_returns} Function to calculate returns on multiple tracks based on temporal and spatial restrictions.
}
