% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_calculate-return-intensity.R
\name{calc_spint_track}
\alias{calc_spint_track}
\title{Calculate Spatial Intensity of Returns (Single Track)}
\usage{
calc_spint_track(r, dist, lag)
}
\arguments{
\item{r}{Data frame of returns obtained from
\code{calc_returns()} for one individual.}

\item{dist}{Numeric. Distance used to calculate returns. Unit should
match the unit of the coordinates: m if UTM, degrees if lat/long.
#' @param lag Numeric. Temporal lag of returns (unit is locations, e.g., 42 for
one week at 6h resolution). Must match one of the lags used in
\code{calc_returns()}.}
}
\value{
A vector with as many elements as the number of locations that were
revisited at least once. Values indicate the number of returns to each
of those revisited locations.
}
\description{
Function to calculate spatial intensity of returns on a single scenario/ID
combination. Called by \code{calc_spint()}.
}
