% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_create-scenarios.R
\name{create_scenarios_ccrw}
\alias{create_scenarios_ccrw}
\title{Create Constrained Correlated Random Walk Scenario(s) (CCRW)}
\usage{
create_scenarios_ccrw(rho, boundary_size)
}
\arguments{
\item{rho}{Numeric. Movement autocorrelation parameter}

\item{boundary_size}{Size of the boundary in m. Numeric value}
}
\value{
Returns a data frame of all the possible Constrained Correlated Random Walk Scenario(s) based on rho and boundary size values. Na is applied to habitat_effect, beta, and lands.
}
\description{
\code{create_scenarios_ccrw} generates Constrained Correlated Random Walk scenario(s).
}
