% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_create-scenarios.R
\name{create_scenarios_crw}
\alias{create_scenarios_crw}
\title{Create Correlated Random Walk Scenario(s) (CRW)}
\usage{
create_scenarios_crw(rho)
}
\arguments{
\item{rho}{Numeric. Movement autocorrelation parameter}
}
\value{
Returns a data frame of all the possible Correlated Random Walk Scenario(s) based on rho values. Na is applied to Boundary_size, habitat_effect, beta, and lands.
}
\description{
\code{create_scenarios_crw} generates Correlated Random Walk scenario(s).
}
