% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_create-scenarios.R
\name{create_scenarios_mcrw}
\alias{create_scenarios_mcrw}
\title{Create Mixture Correlated Random Walk Scenario(s) (MCRW)}
\usage{
create_scenarios_mcrw(rho, habitat_effect, lands)
}
\arguments{
\item{rho}{Numeric. Movement autocorrelation parameter}

\item{habitat_effect}{Numeric. Strength of habitat response, values between XX-XX.}

\item{lands}{\code{raster} of habitat quality (landscape)}
}
\value{
Returns a data frame of all the possible Mixture Correlated Random Walk Scenario(s) based on rho, habitat effect, and lands values. Na is applied to boundary size and beta.
}
\description{
\code{create_scenarios_mcrw} generates Mixture Correlated Random Walk scenario(s).
}
