% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_get-neighborhoods.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get Neighborhoods}
\usage{
get_neighbors(lands, prange)
}
\arguments{
\item{lands}{Raster of habitat quality.}

\item{prange}{Numeric. Perceptual range radius. Units match the
units of the coordinates: m if UTM, degrees if lat/long.}
}
\value{
A list with one element per raster cell. Each element contains the
identifiers of neighboring cells within the specified perceptual range.
}
\description{
Function to create a map of neighboring cells in a raster within a
user-defined perceptual range distance of each cell.
}
