% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-bcrw.R
\name{sim_bcrw}
\alias{sim_bcrw}
\title{Simulate Biased Correlated Random Walk (BCRW)}
\usage{
sim_bcrw(
  n_steps,
  sl_par,
  rho,
  start_loc,
  beta,
  prange,
  lands,
  neighbors,
  jitter,
  scenario_id = NA,
  lands_name = NA
)
}
\arguments{
\item{n_steps}{Numeric. Number of steps to simulate}

\item{sl_par}{Vector of length 2 containing the shape and scale of the step
length distribution (Weibull)}

\item{rho}{Numeric. Movement autocorrelation parameter}

\item{start_loc}{Data frame with a single row containing coordinates of the
starting location (columns \code{x} and \code{y}). 0, 0 if unprojected.}

\item{beta}{Numeric. Value of bias}

\item{prange}{Numeric. Perceptual range radius. Units match the
units of the coordinates: m if UTM, degrees if lat/long.}

\item{lands}{Raster of habitat quality.}

\item{neighbors}{List of cell neighborhoods created by \code{get_neighbors}.
#' @param jitter Logical. If \code{TRUE}, starting locations will be jittered
by 20 km in any direction for spatially-explicit models (MCRW and BCRW).}

\item{scenario_id}{Character string. Optional ID of the current scenario.
Automatically assigned when using simulate_tracks().}

\item{lands_name}{Path to the landscape file.}
}
\value{
Returns a data frame containing the coordinates of each location
along the simulated track and the associated simulation parameter values as
specified in \code{create_scenarios_bcrw}.
}
\description{
Function to simulate a movement track under a Biased Correlated Random Walk
model. Called as needed by \code{simulate_tracks}.
}
