% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-ccrw.R
\name{sim_ccrw}
\alias{sim_ccrw}
\title{Simulate Constrained Correlated Random Walk (CCRW)}
\usage{
sim_ccrw(
  n_steps,
  sl_par,
  rho,
  boundary_size,
  start_loc = data.frame(x = 0, y = 0),
  scenario_id = NA
)
}
\arguments{
\item{n_steps}{Numeric. Number of steps to simulate}

\item{sl_par}{Vector of length 2 containing the shape and scale of the step
length distribution (Weibull)}

\item{rho}{Numeric. Movement autocorrelation parameter}

\item{boundary_size}{Size of the boundary in m. Numeric value}

\item{start_loc}{Data frame with coordinates of the starting location
(columns named x and y). 0, 0 if unprotected.}

\item{scenario_id}{Character string. Optional ID of the current scenario. Automatically assigned
when using simulate_tracks().}
}
\value{
Returns a data frame containing the coordinates of each location along the
simulated track and the associated simulation parameter values.
}
\description{
\code{sim_ccrw} simulates a movement track with a Constrained Correlated Random Walk model.
}
\details{
Insert description
}
