% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-crw.R
\name{sim_crw}
\alias{sim_crw}
\title{Simulate Correlated Random Walk (CRW)}
\usage{
sim_crw(
  n_steps,
  sl_par,
  rho,
  start_loc = data.frame(x = 0, y = 0),
  scenario_id = NA
)
}
\arguments{
\item{n_steps}{Numeric. Number of steps to simulate}

\item{sl_par}{Vector of length 2 containing the shape and scale of the step
length distribution (Weibull)}

\item{rho}{Numeric. Movement autocorrelation parameter}

\item{start_loc}{Data frame with a single row containing coordinates of the
starting location (columns \code{x} and \code{y}). 0, 0 if unprojected.}

\item{scenario_id}{Character string. Optional ID of the current scenario.
Automatically assigned when using simulate_tracks().}
}
\value{
A data frame containing the coordinates of each location along the
simulated track and the associated simulation parameter values as specified
in \code{create_scenarios_crw}.
}
\description{
Function to simulate a movement track under a Correlated Random Walk model.
Called as needed by \code{simulate_tracks}.
}
