% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-landscapes.R
\name{sim_land}
\alias{sim_land}
\title{Generate simulated landscapes}
\usage{
sim_land(res, ext, orig, contagion, constancy)
}
\arguments{
\item{res}{Numeric. Resolution of the raster (units match the units of the
coordinates, m if UTM and degrees if lat/long)}

\item{ext}{Numeric. Raster extent (units match the units of the coordinates,
m if UTM and degrees if lat/long)}

\item{orig}{Named vector of x and y coordinates of the center of the raster.}

\item{contagion}{Numeric vector (any length) of contagion values to simulate.}

\item{constancy}{Numeric vector (any length) of constancy values to simulate.}
}
\value{
\code{Raster} or \code{RasterStack} of simulated landscapes. As many
as \code{length(contagion) * length(constancy)}.
}
\description{
Function to generate one or a set of simulated landscapes of user-specified
extent, resolution, origin, and spatial properties (contagion and constancy)
}
