% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-mcrw.R
\name{sim_mcrw}
\alias{sim_mcrw}
\title{Simulate Mixture Correlated Random Walk (MCRW)}
\usage{
sim_mcrw(
  n_steps,
  sl_par,
  rho,
  start_loc,
  habitat_effect,
  lands,
  scenario_id = NA,
  lands_name = NA
)
}
\arguments{
\item{n_steps}{Numeric. Number of steps to simulate}

\item{sl_par}{Vector of length 2 containing the shape and scale of the step
length distribution (Weibull)}

\item{rho}{Numeric. Movement autocorrelation parameter}

\item{start_loc}{Data frame with coordinates of the starting location
(columns named x and y). 0, 0 if unprotected.}

\item{habitat_effect}{Numeric. Strength of habitat response, values between XX-XX.}

\item{lands}{\code{raster} of habitat quality (landscape)}

\item{scenario_id}{Character string. Optional ID of the current scenario. Automatically assigned
when using simulate_tracks().}

\item{lands_name}{Path to the landscape file}
}
\value{
Returns a data frame containing the coordinates of each location along the
simulated track and the associated simulation parameter values.
}
\description{
\code{sim_mcrw} simulates a movement track with a Mixture Correlated Random Walk model.
}
