% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-tracks.R
\name{simulate_tracks}
\alias{simulate_tracks}
\title{Simulate movement tracks}
\usage{
simulate_tracks(
  start_loc = data.frame(x = 0, y = 0),
  sl_par = c(NA, NA),
  n_steps = NA,
  n_tracks = NA,
  rho = NA,
  boundary_size = NA,
  lands = NA,
  habitat_effect = NA,
  beta = NA,
  prange = NA,
  neighbors = NA,
  model = "all"
)
}
\arguments{
\item{start_loc}{Data frame with coordinates of the starting location
(columns named x and y)}

\item{sl_par}{Vector of length 2 containing the shape and scale of the step
length distribution (Weibull)}

\item{n_steps}{Numeric. Number of steps to simulate}

\item{rho}{Numeric. Movement autocorrelation parameter}
}
\description{
\code{simulate_tracks} is a wrapper function to simulate tracks from multiple
movement models.
}
