% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_create-scenarios.R
\name{create_scenarios_bcrw}
\alias{create_scenarios_bcrw}
\title{Create Biased Correlated Random Walk Scenario(s) (BCRW)}
\usage{
create_scenarios_bcrw(rho, beta, lands)
}
\arguments{
\item{rho}{Numeric. Movement autocorrelation parameter}

\item{beta}{Numeric. Bias applied to the Biased Correlated Random Walk}

\item{lands}{\code{Raster} of habitat quality (landscape)}
}
\value{
Returns a data frame of all the possible Biased Correlated Random Walk Scenario(s) based on rho, beta, and lands values. Na is applied to boundary size and habitat effect.
}
\description{
\code{create_scenarios_mcrw} generates Biased Correlated Random Walk scenario(s).
}
