% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_calc-coords-from-step.R
\name{end_coords}
\alias{end_coords}
\title{Calculate End Point Coordinates}
\usage{
end_coords(step, angle, start_x, start_y)
}
\arguments{
\item{step}{Numeric. The step length in meters (m)}

\item{angle}{Numeric. The turning angle in degrees}

\item{start_x}{Numeric. The X coordinate of the start location (UTM)}

\item{start_y}{Numeric. The Y coordinate of the start location (UTM)}
}
\value{
A data frame containing the end points coordinates from the step and turning angle input.
}
\description{
Function to calculate end point coordinates from step and turning angle
}
\details{
Code adapted from https://rdrr.io/github/ptompalski/UsefulRFunctions/src/R/circular2xy.R
}
