% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_get-neighborhoods.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get Neighborhoods}
\usage{
get_neighbors(lands, prange)
}
\arguments{
\item{lands}{\code{raster} of habitat quality (landscape)}

\item{prange}{Numeric. Perceptual range radius (m) for BCRW.}
}
\value{
Returns a list of the number of neighbors within the perceptual range for each cell in a landscape
}
\description{
\code{get_neighbors} creates a vector of neighbors in the perceptual range for each cell in a landscape.
}
