% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-bcrw.R
\name{sim_bcrw}
\alias{sim_bcrw}
\title{Simulate Biased Correlated Random Walk (BCRW)}
\usage{
sim_bcrw(
  n_steps,
  sl_par,
  rho,
  start_loc,
  beta,
  prange,
  lands,
  neighbors,
  scenario_id = NA,
  lands_name = NA
)
}
\arguments{
\item{n_steps}{Numeric. Number of steps to simulate}

\item{sl_par}{Vector of length 2 containing the shape and scale of the step
length distribution (Weibull)}

\item{rho}{Numeric. Movement autocorrelation parameter}

\item{start_loc}{Data frame with coordinates of the starting location
(columns named x and y). 0, 0 if unprotected.}

\item{beta}{Numeric. Bias applied to the Biased Correlated Random Walk}

\item{prange}{Numeric. Perceptual range radius (m) for BCRW.}

\item{lands}{\code{raster} of habitat quality (landscape)}

\item{neighbors}{\code{raster} of the distance to neighboring cells (neighborhoods)}

\item{scenario_id}{Character string. Optional ID of the current scenario. Automatically assigned
when using simulate_tracks().}

\item{lands_name}{Path to the landscape file}
}
\value{
Returns a data frame containing the coordinates of each location along the
simulated track and the associated simulation parameter values.
}
\description{
\code{sim_bcrw} simulates a movement track with a Biased Correlated Random Walk model.
}
