% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-ccrw.R
\name{sim_ccrw}
\alias{sim_ccrw}
\title{Simulate Constrained Correlated Random Walk (CCRW)}
\usage{
sim_ccrw(id, n_steps, sl_par, rho, start_loc, boundary_size)
}
\arguments{
\item{id}{Numeric. Current individual ID}

\item{n_steps}{Numeric. Number of steps to simulate}

\item{sl_par}{Vector of length 2 containing the shape and scale of the step
length distribution (Weibull)}

\item{rho}{Numeric. Movement autocorrelation parameter}

\item{start_loc}{Data frame with coordinates of the starting location
(columns named x and y)}

\item{boundary_size}{Size of the boundary in m. Numeric value}
}
\description{
Description
}
