% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-crw.R
\name{sim_crw}
\alias{sim_crw}
\title{Simulate Correlated Random Walk (CRW)}
\usage{
sim_crw(id, n_steps, sl_par, rho, start_loc)
}
\arguments{
\item{id}{Numeric. Current individual ID}

\item{n_steps}{Numeric. Number of steps to simulate}

\item{sl_par}{Vector of length 2 containing the shape and scale of the step
length distribution (Weibull)}

\item{rho}{Numeric. Movement autocorrelation parameter}

\item{start_loc}{Data frame with coordinates of the starting location
(columns named x and y)}
}
\value{
A data frame containing the coordinates of each location along the
simulated track and the associated simulation parameter values.
}
\description{
Simulate a movement track with a Correlated Random Walk model.
}
