% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-landscapes.R
\name{sim_land}
\alias{sim_land}
\title{Function to generate simulated landscapes}
\usage{
sim_land(res, ext, orig, contagion, constancy)
}
\arguments{
\item{res}{Resolution of the raster in m. Numeric value.}

\item{ext}{Extent of the raster in m. Numeric value.}

\item{orig}{Named vector of x and y coordinates of the center of the raster.}

\item{contagion}{Vector of contagion values to simulate.}

\item{constancy}{Vector of constancy values to simulate.}
}
\value{
\code{Raster} or \code{RasterStack} of simulated landscapes.
}
\description{
\code{sim_land} generates a simulated landscape of user-specified ext,
resolution, origin, and spatial properties (contagion and constancy)
}
