% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-tracks.R
\name{simulate_tracks}
\alias{simulate_tracks}
\title{Simulate movement tracks}
\usage{
simulate_tracks(
  scenarios,
  n_steps,
  n_tracks,
  start_loc,
  sl_par,
  prange = NA,
  jitter = TRUE,
  neighbors = NA
)
}
\arguments{
\item{scenarios}{Data frame defining parameter values for each scenario as
formatted by \code{create_scenarios_crw()}, \code{create_scenarios_ccrw()},
etc.}

\item{n_steps}{Numeric. Number of steps to simulate}

\item{n_tracks}{Numeric. Number of tracks to simulate.}

\item{start_loc}{Data frame with a single row containing coordinates of the
starting location (columns \code{x} and \code{y}). 0, 0 if unprojected.}

\item{sl_par}{Vector of length 2 containing the shape and scale of the step
length distribution (Weibull)}

\item{prange}{Numeric. Perceptual range radius. Units match the
units of the coordinates: m if UTM, degrees if lat/long.}

\item{neighbors}{List of cell neighborhoods created by \code{get_neighbors}.
#' @param jitter Logical. If \code{TRUE}, starting locations will be jittered
by 20 km in any direction for spatially-explicit models (MCRW and BCRW).}
}
\value{
Returns a data frame containing the coordinates of each location
along the simulated tracks and the associated simulation parameter values as
specified in \code{scenarios}. Each track is assigned an individual \code{id}
under each scenario.
}
\description{
Wrapper function to simulate tracks under multiple scenarios/movement models.
}
