% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity_funs_simulate-tracks.R
\name{simulate_tracks}
\alias{simulate_tracks}
\title{Simulate movement tracks}
\usage{
simulate_tracks(
  scenarios,
  n_steps,
  n_tracks,
  start_loc,
  sl_par,
  prange = NA,
  neighbors = NA
)
}
\arguments{
\item{scenarios}{Data frame. Output of the create_scenarios() functions (one or multiple rbind together)}

\item{n_steps}{Numeric. Number of steps to simulate}

\item{n_tracks}{Numeric. Number of tracks to simulate.}

\item{start_loc}{Data frame with coordinates of the starting location
(columns named x and y). 0, 0 if unprotected.}

\item{sl_par}{Vector of length 2 containing the shape and scale of the step
length distribution (Weibull)}

\item{prange}{Numeric. Perceptual range radius (m) for BCRW.}

\item{neighbors}{\code{raster} of the distance to neighboring cells (neighborhoods)}
}
\value{
Returns a data frame containing the coordinates of each location along the
simulated track of multiple scenarios (including multiple movement models) and the associated simulation parameter values.
}
\description{
\code{simulate_tracks} is a wrapper function to simulate multiple tracks for
multiple scenarios (which can include multiple movement models).
}
